feat: use drawer instead of modal dialog
